Added function to registration.py
